refactor(spdx-util): Split out document-related code to a separate module
Further modularize the code base to have more fine-grained control over
dependencies.
