+++ plugins/package-managers/spdx/build.gradle.kts
@@ -29,6 +29,7 @@ dependencies {
     implementation(projects.downloader)
     implementation(projects.utils.commonUtils)
     implementation(projects.utils.ortUtils)
+    implementation(projects.utils.spdxDocument)
     implementation(projects.utils.spdxUtils)
 
     implementation(libs.jackson.core)
