+++ utils/spdx-document/src/main/kotlin/model/SpdxAnnotation.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.utils.spdx.model
+package org.ossreviewtoolkit.utils.spdxdocument.model
 
 import java.time.Instant
