+++ utils/spdx-document/src/main/kotlin/model/SpdxExternalDocumentReference.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.utils.spdx.model
+package org.ossreviewtoolkit.utils.spdxdocument.model
 
 import org.ossreviewtoolkit.utils.spdx.SpdxConstants
