+++ utils/spdx-document/src/test/assets/spdx-spec-examples/SPDXJSONExample-v2.2-no-ranges.spdx.json
