+++ utils/spdx-document/src/test/assets/spdx-spec-examples/SPDXYAMLExample-2.2-no-ranges.spdx.yaml
